safari-600.7-branch/LayoutTests/platform/ios-sim/fast/canvas/canvas-measureText-ideographicSpace-expected.txt

FAIL spanWidth should be 43.34765625. Was 44.
PASS successfullyParsed is true

TEST COMPLETE
 
